+++ pkg/providers/dns/resource_akamai_dns_zone.go
@@ -186,19 +186,9 @@ func resourceDNSv2Zone(config dnsZoneResourceConfig) *schema.Resource {
 				},
 			},
 			"multi_provider_dnssec": {
-				Type:        schema.TypeList,
+				Type:        schema.TypeBool,
 				Optional:    true,
-				Description: "Multi-signer DNSSEC properties.",
-				MaxItems:    1,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"enabled": {
-							Type:        schema.TypeBool,
-							Optional:    true,
-							Description: "Enables multi-signer DNSSEC for the zone.",
-						},
-					},
-				},
+				Description: "Enables multi-signer DNSSEC for the zone.",
 			},
 			"version_id": {
 				Type:     schema.TypeString,
@@ -499,7 +489,6 @@ func resourceDNSv2ZoneUpdate(ctx context.Context, d *schema.ResourceData, m inte
 	zoneCreate.EndCustomerID = zone.EndCustomerID
 	zoneCreate.ContractID = zone.ContractID
 	zoneCreate.TSIGKey = zone.TSIGKey
-	zoneCreate.MultiProviderDNSSEC = zone.MultiProviderDNSSEC
 	if err := populateDNSv2ZoneObject(d, zoneCreate, logger); err != nil {
 		return diag.FromErr(err)
 	}
@@ -717,15 +706,11 @@ func populateDNSv2ZoneState(d *schema.ResourceData, zoneresp *dns.GetZoneRespons
 		return fmt.Errorf("%w: %s", tf.ErrValueSet, err.Error())
 	}
 
-	multiProviderDnssecListNew := make([]interface{}, 0)
+	multiProviderDnssecEnabled := false
 	if zoneresp.MultiProviderDNSSEC != nil {
-		multiProviderDnssecNew := map[string]interface{}{
-			"enabled": zoneresp.MultiProviderDNSSEC.Enabled,
-		}
-		multiProviderDnssecListNew = append(multiProviderDnssecListNew, multiProviderDnssecNew)
+		multiProviderDnssecEnabled = zoneresp.MultiProviderDNSSEC.Enabled
 	}
-
-	if err := d.Set("multi_provider_dnssec", multiProviderDnssecListNew); err != nil {
+	if err := d.Set("multi_provider_dnssec", multiProviderDnssecEnabled); err != nil {
 		return fmt.Errorf("%w: %s", tf.ErrValueSet, err.Error())
 	}
 
@@ -828,18 +813,12 @@ func populateDNSv2ZoneObject(d *schema.ResourceData, zone *dns.ZoneCreate, logge
 		}
 	}
 
-	multiProviderDnssec, err := tf.GetListValue("multi_provider_dnssec", d)
+	multiProviderDnssec, err := tf.GetBoolValue("multi_provider_dnssec", d)
 	if err != nil && !errors.Is(err, tf.ErrNotFound) {
 		return err
 	}
-	if (err == nil || d.HasChange("multi_provider_dnssec")) && len(multiProviderDnssec) > 0 {
-		multiProviderDnssecMap, ok := multiProviderDnssec[0].(map[string]interface{})
-		if !ok {
-			return fmt.Errorf("'multi_provider_dnssec' entry is of invalid type; should be 'map[string]interface{}'")
-		}
-		zone.MultiProviderDNSSEC = &dns.MultiProviderDNSSEC{
-			Enabled: multiProviderDnssecMap["enabled"].(bool),
-		}
+	zone.MultiProviderDNSSEC = &dns.MultiProviderDNSSEC{
+		Enabled: multiProviderDnssec,
 	}
 
 	TSIGKey, err := tf.GetListValue("tsig_key", d)
@@ -892,7 +871,7 @@ func checkDNSv2Zone(d tf.ResourceDataFetcher) error {
 	if err != nil && !errors.Is(err, tf.ErrNotFound) {
 		return err
 	}
-	multiProviderDnssec, err := tf.GetListValue("multi_provider_dnssec", d)
+	multiProviderDnssec, err := tf.GetBoolValue("multi_provider_dnssec", d)
 	if err != nil && !errors.Is(err, tf.ErrNotFound) {
 		return err
 	}
@@ -913,11 +892,8 @@ func checkDNSv2Zone(d tf.ResourceDataFetcher) error {
 	if signandserve && ztype == "ALIAS" {
 		return fmt.Errorf("sign_and_serve is not valid in %s zone %s configuration", ztype, zone)
 	}
-	if len(multiProviderDnssec) > 0 {
-		multiProviderDnssecMap, ok := multiProviderDnssec[0].(map[string]interface{})
-		if ok && multiProviderDnssecMap["enabled"].(bool) && !signandserve {
-			return fmt.Errorf("multi_provider_dnssec.enabled requires sign_and_serve to be true in zone %s configuration", zone)
-		}
+	if multiProviderDnssec && !signandserve {
+		return fmt.Errorf("multi_provider_dnssec requires sign_and_serve to be true in zone %s configuration", zone)
 	}
 	if ztype != "SECONDARY" && len(tsig) > 0 {
 		return fmt.Errorf("tsig_key can not be populated in %s zone %s configuration", ztype, zone)