+++ pkg/providers/dns/testdata/TestResDnsZone/create_multisigner.tf
@@ -3,14 +3,11 @@ provider "akamai" {
 }
 
 resource "akamai_dns_zone" "multi_signer_test_zone" {
-  contract       = "ctr1"
-  zone           = "multisignerexampleterraform.io"
-  type           = "primary"
-  comment        = "This is a test zone with multi-signer DNSSEC"
-  sign_and_serve = true
-  group          = "grp1"
-
-  multi_provider_dnssec {
-    enabled = true
-  }
+  contract              = "ctr1"
+  zone                  = "multisignerexampleterraform.io"
+  type                  = "primary"
+  comment               = "This is a test zone with multi-signer DNSSEC"
+  sign_and_serve        = true
+  group                 = "grp1"
+  multi_provider_dnssec = true
 }